refactor(parser): spin off parser into a separate module
This refactoring spins off the groovy-parser into a separate, lsp-agnostic module.

The groovy-parser module no longer depends on lsp4j. It provides its own Position and Range types, and its own Result and Error types.

The groovy-lsp module now uses the groovy-parser module for all parsing and AST-related tasks. This improves separation of concerns and will make it easier to maintain and extend the parser in the future.